+++ tests/Unit/Http/SymfonyResponseTaggerTest.php
@@ -11,12 +11,8 @@
 
 namespace FOS\HttpCacheBundle\Tests\Unit;
 
-use FOS\HttpCache\ProxyClient\HttpDispatcher;
-use FOS\HttpCache\ProxyClient\Invalidation\TagCapable;
 use FOS\HttpCache\ProxyClient\ProxyClient;
-use FOS\HttpCache\ProxyClient\Varnish;
 use FOS\HttpCacheBundle\Http\SymfonyResponseTagger;
-use Mockery\MockInterface;
 use Symfony\Component\HttpFoundation\Response;
 
 class SymfonyResponseTaggerTest extends \PHPUnit_Framework_TestCase
@@ -30,29 +26,26 @@ class SymfonyResponseTaggerTest extends \PHPUnit_Framework_TestCase
 
     public function testTagResponse()
     {
-        /** @var TagCapable|MockInterface $tag */
-        $tag = new Varnish(\Mockery::mock(HttpDispatcher::class));
-
         $tags1 = ['post-1', 'posts'];
         $tags2 = ['post-2'];
         $tags3 = ['different'];
 
-        $symfonyResponseTagger1 = new SymfonyResponseTagger($tag);
+        $symfonyResponseTagger1 = new SymfonyResponseTagger();
         $response = new Response();
-        $response->headers->set($tag->getTagsHeaderName(), '');
+        $response->headers->set('X-Cache-Tags', '');
         $symfonyResponseTagger1->addTags($tags1);
         $symfonyResponseTagger1->tagSymfonyResponse($response);
-        $this->assertTrue($response->headers->has($tag->getTagsHeaderName()));
-        $this->assertEquals(implode(',', $tags1), $response->headers->get($tag->getTagsHeaderName()));
+        $this->assertTrue($response->headers->has('X-Cache-Tags'));
+        $this->assertEquals(implode(',', $tags1), $response->headers->get('X-Cache-Tags'));
 
-        $symfonyResponseTagger2 = new SymfonyResponseTagger($tag);
+        $symfonyResponseTagger2 = new SymfonyResponseTagger();
         $symfonyResponseTagger2->addTags($tags2);
         $symfonyResponseTagger2->tagSymfonyResponse($response);
-        $this->assertEquals(implode(',', array_merge($tags2, $tags1)), $response->headers->get($tag->getTagsHeaderName()));
+        $this->assertEquals(implode(',', array_merge($tags2, $tags1)), $response->headers->get('X-Cache-Tags'));
 
-        $symfonyResponseTagger3 = new SymfonyResponseTagger($tag);
+        $symfonyResponseTagger3 = new SymfonyResponseTagger();
         $symfonyResponseTagger3->addTags($tags3);
         $symfonyResponseTagger3->tagSymfonyResponse($response, true);
-        $this->assertEquals(implode(',', $tags3), $response->headers->get($tag->getTagsHeaderName()));
+        $this->assertEquals(implode(',', $tags3), $response->headers->get('X-Cache-Tags'));
     }
 }